+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/FieldsManager.java
@@ -311,7 +311,10 @@ public class FieldsManager implements HeadingListener {
 
     private long getVisibleFieldCount() {
         final List<Field> fields = tableSettings.getFields();
-        return fields.stream().filter(Field::isVisible).count();
+        return fields.stream()
+                .filter(field -> !field.isSpecial())
+                .filter(Field::isVisible)
+                .count();
     }
 
     private Field getField(final int colIndex) {