+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/text/TextPresenter.java
@@ -292,7 +292,12 @@ public class TextPresenter extends AbstractComponentPresenter<TextPresenter.Text
 
                         // If we have a source highlight then use it.
                         if (highlight != null) {
-                            currentPageRange = new OffsetRange<>(highlight.getFrom().getLineNo() - 1L, (long) highlight.getTo().getLineNo() - highlight.getFrom().getLineNo());
+                            // lines 2=>3 means lines 2 & 3, lines 4=>4 means line 4
+                            // -1 to offset to make zero based
+                            // +1 to length to make inclusive
+                            currentPageRange = new OffsetRange<>(
+                                    highlight.getFrom().getLineNo() - 1L,
+                                    (long) highlight.getTo().getLineNo() - highlight.getFrom().getLineNo() + 1);
                         } else {
                             currentPageRange = new OffsetRange<>(sourceLocation.getRecordNo() - 1L, 1L);
                         }
@@ -407,11 +412,17 @@ public class TextPresenter extends AbstractComponentPresenter<TextPresenter.Text
         super.read(componentConfig);
         textSettings = getSettings();
 
-        if (textSettings.getStreamIdField() == null) {
-            textSettings.setStreamIdField(new Field(IndexConstants.STREAM_ID));
+        // special field names have changed from EventId to __event_id__ so we need to deal
+        // with those and replace them, also rebuild existing special fields just in case
+        if (textSettings.getStreamIdField() == null
+                || IndexConstants.STREAM_ID.equals(textSettings.getStreamIdField().getName())
+                || textSettings.getStreamIdField().isSpecial()) {
+            textSettings.setStreamIdField(TablePresenter.buildSpecialField(IndexConstants.STREAM_ID));
         }
-        if (textSettings.getRecordNoField() == null) {
-            textSettings.setRecordNoField(new Field(IndexConstants.EVENT_ID));
+        if (textSettings.getRecordNoField() == null
+                || IndexConstants.EVENT_ID.equals(textSettings.getStreamIdField().getName())
+                || textSettings.getRecordNoField().isSpecial()) {
+            textSettings.setRecordNoField(TablePresenter.buildSpecialField(IndexConstants.EVENT_ID));
         }
     }