+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Change names of hidden special table columns on dashboards to avoid name clashes.
+
+* Make dashboard table settings popup bigger to accomodate the conditional formatting.
+
+* Added logic to rename conditional formatting term fields on a column rename.
+
+* Added logic to prevent renaming a column to an existing name.
+
+* Issue **#1872** : Partially fixed to show the 1st 1k chars of the single line raw source. Full fix will come in v7.
+
+* Issue **#1874** : Fixed dashboard tables not showing data if the stream id column is present.
+
 * Issue **#1868** : Stop `Stream not found with id=nnn` errors during searching.